Added 2 new convars, an icon for the editor, and fixed issues with the existing code
Added a way to set r_radiosity for all players, as this seems to help a bit with static props and dynamic props being lit when they shouldn't be. Also helps with out-door lighting not working right.
Added a checkbox to set the colour of non-player spawned ropes to black. (Ropes have similar issues to other things when it comes to not updating their lighting)

Removed a few weird things I did instead of using the type() function/tostring and tonumber.

Added a system to choose which light style to set lamps/dynamic lights to, avoiding ones that are used by existing dynamic lights on the map.

Also removed some commented-out code.