Fixed editor, Greatly improved water-darkening, improved misc pieces of code.
Replaced the horrible code in the editor with a version using NetworkVarNotify (what I should've been using in the first place).

Replaced the crappy transparent rectangle for water-darkening with code that accesses the water's material, and changes its fog through that.

Got rid of an un-used network string.

Removed the code meant to wake prop_physics in the amblight/sunlight functions, as it doesn't work, and wouldn't help in the way I thought it would even if it did.

 Replaced the conversion from slider-value to pattern character in amblight/sunlight with an ascii conversion.