Fixed some small issues
GetConvar doesn't always exist in one of the context's it's used in, so the code accounts for that now.
ForceDisableSkybox could be ignored by setting r_3dsky to any value other than 0 or 1 to get the skybox back, it now checks for values other than "0" instead of checking specifically for "1".