Refactor, QoL, Extra features for static props.
Refactored most of my old code so that it sucks a bit less/is more easily navigable.

Added QoL improvements that have been sitting in the github issues for like a year. Including ConVars resetting after disconnect, Addon cvars resetting on map clean-up (with an option to disable it for obscure cases where people might want to, like in gamemodes that clean the map after every round.), and trying to keep editor entity values consistent with the actual CVars where possible.

Finally found (and implemented) a way of getting rid of the last of the lighting on certain static props (i.e. the buildings in gm_construct's skybox, or the roads/trees on gm_boreas). Also added a way to disable self-illumination on materials used by static props using the same method.

The actual way I'm doing this is incredibly cursed. I'm reading from the map's BSP file directly, finding any model names, filtering out anything that isn't a static prop, setting a temporary entity's model to each found model, and using that entity to grab the model's materials. *Then* I'm giving each material a flat normal map, because that gets rid of the ambient light on static props for some reason.