Bug fix for water darkening
The "default colour" value for map's fog materials was being stored as a reference to the original vector, rather than as a copy. So whenever the original was set to a new value, that new value would be treated as the original, resulting in colour being shifted around a moving point rather than the original and water getting too dark (or bright, but I don't think many people use this addon to make things brighter).