ForceDisableDetailProps, misc small changes/fixes.
Added a convar to set r_drawdetailprops to 0 on all clients, because they tend to respond weirdly to different lighting (either being fully lit or fully unlit regardless of nearby light). I *might* be able to fix this for the billboard grass effects that appear on displacements, but that's something to try another time. I'm pretty sure 3D detail props are unfixable regardless, so even if I do manage that, this convar will still be useful.

Moved the prints in sv_staticproplighting around, as they were printing at the wrong times.

Renamed the menu option for removing particle systems to refer to them as "Particle Emitters" to make things clearer for people with less source-engine knowledge.
Added a value for removing particle systems on the editor entity, because I forgot to do that last time.

Made it so the code responsible for keeping editors and convars synced is only serverside, and not shared.